chore: 공통 DB 스키마 정의 및 스웨거 문서 추가 (#1)
