+++ docker-compose.yml
@@ -23,7 +23,6 @@ services:
       - "5432:5432"
     volumes:
       - pgdata:/var/lib/postgresql/data
-      - ./db/init.sql:/docker-entrypoint-initdb.d/init.sql:ro
     healthcheck:
       test: ["CMD-SHELL", "pg_isready -U ${POSTGRES_USER} -d ${POSTGRES_DB}"]
       interval: 10s
@@ -52,6 +51,11 @@ services:
       AI_MOCK:                    ${AI_MOCK}
     ports:
       - "8080:8080"
+    healthcheck:
+      test: ["CMD-SHELL", "wget -q -O - http://localhost:8080/api/v1/health > /dev/null || exit 1"]
+      interval: 10s
+      timeout: 5s
+      retries: 10
     volumes:
       - shared_data:/data
     depends_on:
@@ -77,6 +81,8 @@ services:
       POSTGRES_PASSWORD:       ${POSTGRES_PASSWORD}
     depends_on:
       db:
+        condition: service_healthy
+      backend:
         condition: service_healthy
     volumes:
       - shared_data:/data
