--- src/main/java/com/campost/backend/common/api/ApiResponse.java
@@ -1,13 +1,0 @@
-package com.campost.backend.common.api;
-
-import java.time.OffsetDateTime;
-
-public record ApiResponse<T>(
-        boolean success,
-        T data,
-        OffsetDateTime timestamp
-) {
-    public static <T> ApiResponse<T> ok(T data) {
-        return new ApiResponse<>(true, data, OffsetDateTime.now());
-    }
-}
