--- src/main/java/com/campost/backend/common/config/CorsConfig.java
@@ -1,34 +1,0 @@
-package com.campost.backend.common.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.lang.NonNull;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import java.util.Arrays;
-
-@Configuration
-public class CorsConfig implements WebMvcConfigurer {
-
-    @Value("${app.cors.allowed-origins:http://localhost:5173,http://localhost:3000}")
-    private String allowedOrigins = "http://localhost:5173,http://localhost:3000";
-
-    @Override
-    public void addCorsMappings(@NonNull CorsRegistry registry) {
-        String[] origins = Arrays.stream(allowedOrigins.split(","))
-                .map(String::trim)
-                .filter(origin -> !origin.isBlank())
-                .toArray(String[]::new);
-
-        if (origins.length == 0) {
-            origins = new String[]{"http://localhost:5173", "http://localhost:3000"};
-        }
-
-        registry.addMapping("/api/**")
-                .allowedOrigins(origins)
-                .allowedMethods("GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS")
-                .allowedHeaders("*")
-                .allowCredentials(true);
-    }
-}
