+++ src/main/java/com/campost/backend/domain/admin/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Admin domain.
+ * <p>사용자 관리, 역할 관리(GUEST/USER/ADMIN), 크롤러 상태 모니터링.</p>
+ */
+package com.campost.backend.domain.admin;
