+++ src/main/java/com/campost/backend/domain/auth/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Auth domain.
+ * <p>회원가입, 학교 이메일 인증, JWT 로그인, 권한 경계 처리.</p>
+ * <p>계정 기본 정보(학과/학년 설정 등) 책임도 이 도메인에서 다룬다.</p>
+ */
+package com.campost.backend.domain.auth;
