+++ src/main/java/com/campost/backend/domain/collect/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Collect domain.
+ * <p>크롤링 파이프라인 상태, 중복 제거, raw 수집/적재, 파서 로그 모니터링.</p>
+ * <p>세부 구현은 query, importer 하위 패키지로 분리한다.</p>
+ */
+package com.campost.backend.domain.collect;
