+++ src/main/java/com/campost/backend/domain/collect/query/dto/CrawlJobDto.java
@@ -1,0 +1,17 @@
+package com.campost.backend.domain.collect.query.dto;
+
+import java.time.OffsetDateTime;
+
+public record CrawlJobDto(
+        long id,
+        long sourceId,
+        OffsetDateTime startedAt,
+        OffsetDateTime finishedAt,
+        String status,
+        int totalFound,
+        int newCount,
+        int skipCount,
+        int failCount,
+        String errorMsg
+) {
+}
