+++ src/main/java/com/campost/backend/domain/collect/query/service/CrawlQueryService.java
@@ -1,0 +1,37 @@
+package com.campost.backend.domain.collect.query.service;
+
+import com.campost.backend.domain.collect.query.dto.CrawlJobDto;
+import com.campost.backend.domain.collect.query.dto.ParseLogDto;
+import com.campost.backend.domain.collect.query.repository.CrawlQueryRepository;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class CrawlQueryService {
+
+    private static final int MAX_LIMIT = 100;
+
+    private final CrawlQueryRepository crawlQueryRepository;
+
+    public CrawlQueryService(CrawlQueryRepository crawlQueryRepository) {
+        this.crawlQueryRepository = crawlQueryRepository;
+    }
+
+    public List<CrawlJobDto> getRecentCrawlJobs(int limit) {
+        int safeLimit = normalizeLimit(limit);
+        return crawlQueryRepository.findRecentCrawlJobs(safeLimit);
+    }
+
+    public List<ParseLogDto> getRecentParseLogs(int limit) {
+        int safeLimit = normalizeLimit(limit);
+        return crawlQueryRepository.findRecentParseLogs(safeLimit);
+    }
+
+    private int normalizeLimit(int limit) {
+        if (limit <= 0) {
+            return 20;
+        }
+        return Math.min(limit, MAX_LIMIT);
+    }
+}
