+++ src/main/java/com/campost/backend/domain/community/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Community domain.
+ * <p>댓글, 대댓글, 좋아요 기능을 담당한다. (후속 구현)</p>
+ */
+package com.campost.backend.domain.community;
