+++ src/main/java/com/campost/backend/domain/hub/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Hub domain.
+ * <p>대시보드, 검색, 스마트 필터(학과/카테고리), 정렬/집계 API.</p>
+ */
+package com.campost.backend.domain.hub;
