+++ src/main/java/com/campost/backend/domain/notification/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Notification domain.
+ * <p>개인화 알림(키워드 알림, 마감 임박 알림) 정책과 전송 이력을 관리한다.</p>
+ */
+package com.campost.backend.domain.notification;
