+++ src/main/java/com/campost/backend/domain/personal/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Personal domain.
+ * <p>북마크/스크랩, 관심 키워드, 개인 보관함, 사용자 맞춤 알림.</p>
+ */
+package com.campost.backend.domain.personal;
