+++ src/main/java/com/campost/backend/domain/post/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Post domain.
+ * <p>공지 목록/상세, 원문 링크, 핵심 정보 카드(마감일/대상/혜택).</p>
+ * <p>세부 구현은 notice 하위 패키지로 분리한다.</p>
+ */
+package com.campost.backend.domain.post;
