+++ src/main/java/com/campost/backend/domain/ugc/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * UGC domain.
+ * <p>제보 등록, 승인 대기, 수정/삭제 요청 처리 기능을 담당한다. (후속 구현)</p>
+ */
+package com.campost.backend.domain.ugc;
