+++ src/main/java/com/campost/backend/global/exception/InvalidTokenException.java
@@ -1,0 +1,11 @@
+package com.campost.backend.global.exception;
+
+public class InvalidTokenException extends RuntimeException {
+    public InvalidTokenException() {
+        super();
+    }
+
+    public InvalidTokenException(String message) {
+        super(message);
+    }
+}
