+++ src/main/java/com/campost/backend/global/exception/TokenExpiredException.java
@@ -1,0 +1,11 @@
+package com.campost.backend.global.exception;
+
+public class TokenExpiredException extends RuntimeException {
+    public TokenExpiredException() {
+        super();
+    }
+
+    public TokenExpiredException(String message) {
+        super(message);
+    }
+}
