+++ src/main/resources/application.yml
@@ -9,12 +9,21 @@ spring:
     username: ${SPRING_DATASOURCE_USERNAME:campost}
     password: ${SPRING_DATASOURCE_PASSWORD:campost1234}
     driver-class-name: org.postgresql.Driver
+  flyway:
+    enabled: true
+    locations: classpath:db/migration
 
 management:
   endpoints:
     web:
       exposure:
         include: health,info
+
+springdoc:
+  api-docs:
+    path: /v3/api-docs
+  swagger-ui:
+    path: /swagger-ui.html
 
 app:
   cors:
